fix(tech-debt): Split FileBrowser into modular package (Issue #113)
Refactors the large 907-line FileBrowser class (45 methods) into a
modular package structure:

- file_browser/navigation.py: FileBrowserNavigation mixin (82 lines)
  - Cursor movement (up/down/top/bottom)
  - Directory navigation (enter/parent)
  - Hidden files toggle

- file_browser/actions.py: FileBrowserActions mixin (553 lines)
  - File save/execute/edit operations
  - Edit mode switching with vim editor
  - Selection mode switching
  - Modal result handling
  - Helper classes (FileEditTextArea, FileSelectionTextArea, FileBrowserVimApp)

- file_browser/view.py: FileBrowserView (338 lines)
  - Main widget combining navigation + actions mixins
  - Compose/mount lifecycle
  - Reactive state management
  - Key event handling

- file_browser/__init__.py: Package exports
  - Backward compatible FileBrowser alias

The original file_browser.py now serves as a thin re-export module
for backward compatibility.
